Update creature parameters and behaviors
Now Veemon will be your slave, so if you see him on the 3rd, don't hesitate to lure him using the whistle or with food to make him your guard. If you don't believe me, try it in your world and you'll see that he will act as a guardian.